#38
- http-server: специфичные константы и утилиты перенесены в пресеты
- http-server: неспецифичные утилиты перенесены в utils/express
