--- src/http-server/constants.ts
@@ -1,4 +1,0 @@
-export const RESPONSE_EVENT = {
-  renderStart: Symbol('render:start'),
-  renderFinish: Symbol('render:finish'),
-} as const;
