--- src/http-server/errors.ts
@@ -1,6 +1,0 @@
-import { ResponseError } from '../http/errors';
-
-/**
- * @deprecated Следует использовать ResponseError.
- */
-export const SSRError = ResponseError;
