+++ src/preset/parts/constants.ts
@@ -1,0 +1,7 @@
+/**
+ * События в процессе ответа на запрос.
+ */
+export const RESPONSE_EVENT_TYPE = {
+  renderStart: Symbol('render:start'),
+  renderFinish: Symbol('render:finish'),
+} as const;
