+++ src/preset/parts/types.ts
@@ -1,4 +1,5 @@
 import type { Provider, Token } from '../../di';
+import type { Request, Response, NextFunction } from 'express';
 
 export interface PresetTuner {
   (tools: { override: <T>(token: Token<T>, provider: Provider<T>) => void }): void;
@@ -12,4 +13,35 @@ export type KnownHttpApiKey = 'ilium' | 'simaV3' | 'simaV4' | 'simaV6';
 export interface StrictMap<Key extends string> {
   /** Возвращает значение по ключу. */
   get(key: Key): string;
+}
+
+/**
+ * Контекст обработчика express.
+ */
+export interface HandlerContext {
+  req: Request;
+  res: Response;
+  next: NextFunction;
+}
+
+/**
+ * Внутренний интерфейс для хранения ссылок на ассеты.
+ */
+export interface PageAssets {
+  js: string;
+  css: string;
+  criticalJs?: string;
+  criticalCss?: string;
+}
+
+/**
+ * Структура ответа от frontend-микросервиса в формате JSON по соглашению.
+ */
+export interface ConventionalJson {
+  markup: string;
+  bundle_js: string;
+  bundle_css: string;
+  critical_js?: string;
+  critical_css?: string;
+  meta?: any;
 }
