+++ src/tokens.ts
@@ -1,11 +1,12 @@
 import { createToken } from './di';
+
+// ВАЖНО: чтобы токенами можно было пользоваться независимо от среды исполнения - импортировать надо только типы
 import type * as express from 'express';
 import type { ConfigSource, BaseConfig } from './config/types';
 import type { Logger } from './log/types';
 import type { Cache } from './cache/types';
 import type { HttpClientFactory } from './http-client/types';
 import type { LogMiddlewareHandlerInit } from './http-client/middleware/log';
-import type { PageAssets, HandlerContext } from './http-server/types';
 import type { SagaExtendedMiddleware } from './utils/redux-saga';
 import type { BridgeClientSide, BridgeServerSide } from './utils/ssr';
 import type { Tracer } from '@opentelemetry/api';
@@ -14,7 +15,7 @@ import type { Resource } from '@opentelemetry/resources';
 import type { ElementType, ReactNode } from 'react';
 
 // @todo придумать как брать не из пресетов:
-import type { StrictMap, KnownHttpApiKey } from './preset/parts/types';
+import type { PageAssets, HandlerContext, StrictMap, KnownHttpApiKey } from './preset/parts/types';
 import type { SpecificExtras } from './preset/node/handler';
 
 export const KnownToken = {
