+++ src/utils/express/index.ts
@@ -1,0 +1,2 @@
+export { healthCheck } from './handler/health-check';
+export { composeMiddleware } from './utils';
